Organize read one tests around relationships
