+++ test/integration/get-one-resource/success.js
@@ -8,7 +8,6 @@ const applyMigrations = require('../../helpers/apply-migrations');
 const seed = require('../../helpers/seed');
 
 const db = getDb();
-const fixturesDirectory = path.join(__dirname, '..', '..', 'fixtures');
 
 describe('Resource GET (one) success', function() {
   // Ensure that the DB connection drops immediately after each test
@@ -25,7 +24,7 @@ describe('Resource GET (one) success', function() {
   describe('when the request succeeds', () => {
     beforeEach((done) => {
       this.options = {
-        resourcesDirectory: path.join(fixturesDirectory, 'kitchen-sink'),
+        resourcesDirectory: path.join(global.fixturesDirectory, 'kitchen-sink'),
         apiVersion: 10
       };
 
@@ -66,7 +65,7 @@ describe('Resource GET (one) success', function() {
   describe('when an empty sparse fieldsets param is specified', () => {
     beforeEach((done) => {
       this.options = {
-        resourcesDirectory: path.join(fixturesDirectory, 'kitchen-sink'),
+        resourcesDirectory: path.join(global.fixturesDirectory, 'kitchen-sink'),
         apiVersion: 16
       };
 
@@ -108,7 +107,7 @@ describe('Resource GET (one) success', function() {
   describe('when the request succeeds with sparse fieldsets', () => {
     beforeEach((done) => {
       this.options = {
-        resourcesDirectory: path.join(fixturesDirectory, 'kitchen-sink'),
+        resourcesDirectory: path.join(global.fixturesDirectory, 'kitchen-sink'),
         apiVersion: 1
       };
 
@@ -138,134 +137,6 @@ describe('Resource GET (one) success', function() {
       request(app(this.options))
         .get('/v1/no_metas/1')
         .query('fields[no_metas]=first_name')
-        .expect(validators.basicValidation)
-        .expect(validators.assertData(expectedData))
-        .expect(validators.assertLinks(expectedLinks))
-        .expect(200)
-        .end(done);
-    });
-  });
-
-  describe('when the request succeeds, with a host relationship', () => {
-    beforeEach((done) => {
-      this.options = {
-        resourcesDirectory: path.join(fixturesDirectory, 'kitchen-sink'),
-        apiVersion: 2
-      };
-
-      const relationGuestSeeds = [
-        {first_name: 'sandwiches'},
-        {first_name: 'what'},
-        {first_name: 'pls'}
-      ];
-
-      const relationSeeds = [
-        {name: 'james', owner_id: '1'}
-      ];
-
-      applyMigrations(this.options)
-        .then(() => seed('relation_guest', relationGuestSeeds))
-        .then(() => seed('relation', relationSeeds))
-        .then(() => done());
-    });
-
-    it('should return a 200 OK, with the resource', (done) => {
-      const expectedData = {
-        type: 'relations',
-        id: '1',
-        attributes: {
-          name: 'james',
-          size: null
-        },
-        relationships: {
-          owner: {
-            data: {
-              id: '1',
-              type: 'relation_guests'
-            },
-            links: {
-              related: '/v2/relations/1/owner',
-              self: '/v2/relations/1/relationships/owner'
-            }
-          }
-        }
-      };
-
-      const expectedLinks = {
-        self: '/v2/relations/1'
-      };
-
-      request(app(this.options))
-        .get('/v2/relations/1')
-        .expect(validators.basicValidation)
-        .expect(validators.assertData(expectedData))
-        .expect(validators.assertLinks(expectedLinks))
-        .expect(200)
-        .end(done);
-    });
-  });
-
-  describe('when the request succeeds, with a guest relationship', () => {
-    beforeEach((done) => {
-      this.options = {
-        resourcesDirectory: path.join(fixturesDirectory, 'kitchen-sink'),
-        apiVersion: 2
-      };
-
-      const relationGuestSeeds = [
-        {first_name: 'sandwiches'},
-        {first_name: 'what'},
-        {first_name: 'pls'}
-      ];
-
-      const relationSeeds = [
-        {name: 'james', owner_id: '1'},
-        {name: 'hungry', owner_id: '3'},
-        {name: 'pizza', owner_id: '1'},
-        {name: 'sammy', owner_id: '2'},
-        {name: 'meow', owner_id: '1'},
-      ];
-
-      applyMigrations(this.options)
-        .then(() => seed('relation_guest', relationGuestSeeds))
-        .then(() => seed('relation', relationSeeds))
-        .then(() => done());
-    });
-
-    it('should return a 200 OK, with the resource and its related items', (done) => {
-      const expectedData = {
-        type: 'relation_guests',
-        id: '1',
-        attributes: {
-          first_name: 'sandwiches',
-          last_name: null
-        },
-        relationships: {
-          pets: {
-            data: [
-              {id: '1', type: 'relation'},
-              {id: '3', type: 'relation'},
-              {id: '5', type: 'relation'},
-            ],
-            links: {
-              self: '/v2/relation_guests/1/relationships/pets',
-              related: '/v2/relation_guests/1/pets'
-            }
-          },
-          one_to_one: {
-            links: {
-              self: '/v2/relation_guests/1/relationships/one_to_one'
-            }
-          }
-        }
-      };
-
-      const expectedLinks = {
-        self: '/v2/relation_guests/1',
-      };
-
-      request(app(this.options))
-        .get('/v2/relation_guests/1')
         .expect(validators.basicValidation)
         .expect(validators.assertData(expectedData))
         .expect(validators.assertLinks(expectedLinks))
